refactor(myProfile/adapter): 리사이클러 뷰 어댑터와 뷰 홀더 파일 분리
