+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/main/myProfile/adapter/ActivitiesAdapter.kt
@@ -1,12 +1,10 @@
 package com.emmsale.presentation.ui.main.myProfile.adapter
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.emmsale.databinding.ItemMyprofileActivitiesBinding
 import com.emmsale.presentation.ui.main.myProfile.uiState.ActivityUiState
 
 class ActivitiesAdapter : ListAdapter<ActivityUiState, ActivityViewHolder>(diffUtil) {
@@ -30,27 +28,6 @@ class ActivitiesAdapter : ListAdapter<ActivityUiState, ActivityViewHolder>(diffU
                 oldItem: ActivityUiState,
                 newItem: ActivityUiState,
             ): Boolean = oldItem == newItem
-        }
-    }
-}
-
-class ActivityViewHolder(
-    private val binding: ItemMyprofileActivitiesBinding
-) : RecyclerView.ViewHolder(binding.root) {
-
-    fun bind(activity: ActivityUiState) {
-        binding.activity = activity
-    }
-
-    companion object {
-        fun create(parent: ViewGroup): ActivityViewHolder {
-            val binding = ItemMyprofileActivitiesBinding.inflate(
-                LayoutInflater.from(parent.context),
-                parent,
-                false
-            )
-
-            return ActivityViewHolder(binding)
         }
     }
 }
